Car Speed Detection
Computes speed of the car by accepting images of the car in two different time frames and finding relative distance between initial and final position of the car. Uses a background image for finding the position of car in 1 image. Time difference between the images and distance of camera form road is required.